tools/TeachingAssistant.ContentPipeline/requirements.txt
# Content Pipeline Dependencies
anthropic>=0.40.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
markdown>=3.5.1
pyyaml>=6.0.1
tiktoken>=0.7.0
numpy>=1.26.0
transformers>=4.40.0
torch>=2.2.0
sentence-transformers>=2.3.0
httpx>=0.27.0
